GH-39212: [Release] Retry download binary subprocess in case of OpenSSL error (#39213)
### Rationale for this change

In case of OpenSSL error when trying to download RC binaries we do not retry and we are facing some issues with artifactory that after some job retries allowed us to successfully download the binaries.

### What changes are included in this PR?

Retry mechanism in case of subprocess error.

### Are these changes tested?

Tested locally verifying wheels and jars for 14.0.2 RC 3.

### Are there any user-facing changes?

No
* Closes: #39212

Lead-authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>