MINOR: [Docs] Add "Field" as a term to "Overview of Arrow Terminology" (#45112)
### Rationale for this change

In the intro docs ("Overview of Arrow Terminology" section) the sentence `[...]A Field includes a field name, a data type, a nullability flag and optional key-value metadata for a specific column in a RecordBatch.[...]` is positioned in a non-consistent way compared to others. This felt a bit confusing in the beginning as I thought, that either the term, which should be described is missing or that the sentence should belong to `Schema`.

<img width="736" alt="image" src="https://github.com/user-attachments/assets/afac4db0-10a8-46e7-a384-80faffad2334" />

(Here you can see that the sentence from above is positioned in a way implying that it also describes an Arrow term on its own)

As `field` also appears in the images at the end of the section I suggest that it should also be treated as a separate term, therefore adding the sub-heading (Initially I thought it should be moved closer to **Schema**, but it feels more consistent to treat it as a separate term).

### What changes are included in this PR?

Added a **Field** sub-heading to "Overview of Arrow Terminology"

### Are these changes tested?

No, as this is only a docs change.

### Are there any user-facing changes?

No, as this is only a docs change.

Authored-by: Tim Grein <grein.tim97@gmail.com>
Signed-off-by: Rossi Sun <zanmato1984@gmail.com>