GH-43352: [Docs][Python] Add all tensor classes documentation
### Rationale for this change

We have added the missing documentation for various tensor classes in PyArrow, specifically `SparseCOOTensor`, `SparseCSRMatrix`, `SparseCSCMatrix`, and `SparseCSFTensor`. This documentation is necessary to provide users with reference information and examples on how to use these classes, as currently, there is no reference documentation or information in the PyArrow user guide regarding these tensor classes.

### What changes are included in this PR?

1. Added docstrings to the `SparseCSRMatrix`, `SparseCSCMatrix`, and `SparseCSFTensor` classes and their methods in the `tensor.pxi` file.
2. Updated the `tables.rst` file to include documentation for `SparseCSFTensor` along with examples.

### Are these changes tested?

No new tests are included in this PR as the changes are purely documentation updates. The existing tests for the tensor classes should cover the functionality.

### Are there any user-facing changes?

Yes, the user-facing change is the addition of documentation for the tensor classes, which will help users understand and utilize these classes more effectively. There are no breaking changes to public APIs.