GH-45092: [C++][Parquet] Add GetReadRanges function to FileReader (#45093)
### Rationale for this change
For some consumers, it is convenient to expose a way to retrieve the necessary byte ranges of a parquet file to read specific column chunks from row groups without having to go through a full `ReadRangeCache`. Ultimately, it's a fairly simple function since we already have all the infrastructure implemented to compute these ranges.

### What changes are included in this PR?
A single function added to `parquet::FileReader` named `GetReadRanges` which computes and retrieves the coalesced read ranges for specified row groups and column indices.

* GitHub Issue: #45092

Authored-by: Matt Topol <zotthewizard@gmail.com>
Signed-off-by: Matt Topol <zotthewizard@gmail.com>