GH-45169: [Python] Adapt to modified pytest ignore collect hook api (#45170)
### Rationale for this change

Prevent pytest from logging a warning during the unit tests and adapt to the future removal of the path argument in the `pytest_ignore_collect` hook.

### What changes are included in this PR?

This changes the deprecated `path` argument in  the pytest `pytest_ignore_collect` hook to `collection_path` which it should be as of pytest 7.0: https://docs.pytest.org/en/latest/deprecations.html#py-path-local-arguments-for-hooks-replaced-with-pathlib-path

### Are these changes tested?

The existing tests keep on running without a deprecation warning being logged for this. 

### Are there any user-facing changes?
 No

* GitHub Issue: #45169

Authored-by: Rob Van Mieghem <robvanmieghem@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>