GH-45283: [C++][Parquet] Omit level histogram when max level is 0 (#45285)
### Rationale for this change

The level histogram of size statistics can be omitted if its max level is 0. We haven't implemented this yet and enforces histogram size to be equal to `max_level + 1`. However, when reading a Parquet file with omitted level histogram, exception will be thrown.

### What changes are included in this PR?

Omit level histogram when max level is 0.

### Are these changes tested?

Yes, a test case has been added to reflect the change.

### Are there any user-facing changes?

No.
* GitHub Issue: #45283

Lead-authored-by: Gang Wu <ustcwg@gmail.com>
Co-authored-by: Antoine Pitrou <pitrou@free.fr>
Signed-off-by: Gang Wu <ustcwg@gmail.com>