GH-44759: [GLib] Add garrow_record_batch_validate() (#45353)
### Rationale for this change

[RecordBatch::Validate](https://arrow.apache.org/docs/cpp/api/table.html#_CPPv4NK5arrow11RecordBatch8ValidateEv) available in the C++ API.
But, GLib doesn't support that method yet.

### What changes are included in this PR?

This PR adds a validation method in the record-batch class.
Before this change, the `Validate()` method was used in the `garrow_record_batch_new` implicitly.
This PR removes it and adds it as a separate method. Users need to call `garrow_record_batch_validate()` explicitly by themselves. This is a backward incompatible change.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

Yes.

**This PR includes breaking changes to public APIs.**

* GitHub Issue: #44759

Authored-by: Hiroyuki Sato <hiroysato@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>