GH-45393: [C++][Compute] Fix wrong decoding for 32-bit column in row table (#45473)
### Rationale for this change

The failure reported in #45393 seems to be caused by a careless parentheses typo introduced in #45336:
https://github.com/apache/arrow/blob/e32f56b478171fc4b53dc2042c4cf5d37c97e351/cpp/src/arrow/compute/row/encode_internal.cc#L281-L282

And unfortunately our `Grouper` UT doesn't have cases covering this particular code path (the questioning code path is only triggered in `Grouper` with very restrictive conditions: the row table is fixed-length, a 32-bit key is encoded after some other keys).

### What changes are included in this PR?

An UT to reproduce the issue and the fix.

### Are these changes tested?

UT included.

### Are there any user-facing changes?

None.

* GitHub Issue: #45393

Authored-by: Rossi Sun <zanmato1984@gmail.com>
Signed-off-by: Rossi Sun <zanmato1984@gmail.com>