GH-45377: [CI][R] Ensure install R on ubuntu-24.04 runner for R nightly build jobs (#45464)
### Rationale for this change

Related to #45377.
As a result of the recent change in the ubuntu-latest GHA runner from ubuntu-22.04 to ubuntu-24.04, it appears that the tools included in the runner from the start have changed and jobs that assume R is already installed will now fail.
Installing R the job should now succeed.

### What changes are included in this PR?

### Are these changes tested?

### Are there any user-facing changes?

* GitHub Issue: #45377

Authored-by: eitsupi <50911393+eitsupi@users.noreply.github.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>