GH-45159: [CI][Integration] Remove substrait consumer-testing integration job (#45463)
### Rationale for this change

The job has been failing for the last two months due to upstream refactoring. I opened an issue upstream but we didn't got any response:
- https://github.com/substrait-io/consumer-testing/issues/196

Based on the following substrait consumer-testing PR there was a big refactor which removed the files that we were invoking on our tests:
- https://github.com/substrait-io/consumer-testing/pull/162

I asked on the corresponding issue ~3 weeks ago for feedback on the removal and as there's no opposition I think we should remove this unmaintained job.

### What changes are included in this PR?

Remove extremely coupled `substrait-io/consumer-testing` job.

### Are these changes tested?

No

### Are there any user-facing changes?

No
* GitHub Issue: #45159

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>