GH-45521: [CI][Dev][R] Install required cyclocomp package to be used with R lintr (#45524)
### Rationale for this change

The linting jobs are failing due to the new version of `lintr` not installing `cyclocomp` anymore.
We use `cyclocomp` but this is not part of the default linters of `lintr` anymore. We should install it individually.

### What changes are included in this PR?

Install `cyclocomp` as part of setting up the linting environment for R on our linting job.
Pin old version of `lintr` for R Windows job as it otherwise fails with a lot of new linter issues.

### Are these changes tested?

Yes via CI.

### Are there any user-facing changes?

No
* GitHub Issue: #45521

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>