GH-45532: [C++] `RunEndEncodedBuilder` should clear dimensions after a `Finish()` call (#45533)
### Rationale for this change

This resolves the issue reported in https://github.com/apache/arrow/issues/45532. Specifically, this updates the `RunEndEncodedBuilder` dimensions after a `Finish()` call and therefore makes the function idempotent.

### What changes are included in this PR?

This updates the `RunEndEncodedBuilder`'s dimensions after a `Finish()` call and adds a test which ensures the function properly resets the builder state and is idempotent.

### Are these changes tested?

Yes

### Are there any user-facing changes?

Yes; this is a bug fix to a user-facing API.

* GitHub Issue: #45532

Lead-authored-by: Lester Fan <lesterfan37@gmail.com>
Co-authored-by: mwish <maplewish117@gmail.com>
Signed-off-by: mwish <maplewish117@gmail.com>