GH-43807: [C++][Python] Add UUID extension type conversion support to/from Parquet (#45866)
### Rationale for this change

After https://github.com/apache/arrow/pull/37298, the UUID canonical
extension type is supported in Arrow C++ and PyArrow; however, it is not
converted to the Parquet UUID type on write and is not inferred when on
Parquet read.

### What changes are included in this PR?

- Infer the Parquet UUID type from the Arrow UUID extension type
- Infer the Arrow UUID type from Parquet UUID when
arrow_extensions_enabled is set (like the JSON extension type)
- Wire up arrow_extensions_enabled to pyarrow and add Python tests

### Are these changes tested?

Yes!

### Are there any user-facing changes?

Yes! (Documentation will be added)
* GitHub Issue: #43807

---------

Co-authored-by: Gang Wu <ustcwg@gmail.com>