GH-46231: [C++][CMake] Fix `arrow_bundled_dependencies` to be externally accessible by FetchContent (#46232)
### Rationale for this change

`arrow_bundled_dependencies` is an IMPORTED library which links to `arrow_bundled_dependencies_merge`. When arrow-cpp is used externally as a third-party dependency via CMake `FetchContent`, target `arrow_bundled_dependencies` is not accessible by the parent project (but `arrow_bundled_dependencies_merge` is). It is difficult for the parent project to locate the built product of `arrow_bundled_dependencies_merge` and install it.

### What changes are included in this PR?

Change `IMPORTED` library `arrow_bundled_dependencies` to be `GLOBAL`.

### Are these changes tested?

- Pass all CIs.
- Verified to work by https://github.com/apache/iceberg-cpp/pull/89

### Are there any user-facing changes?

I don't think so.

* GitHub Issue: #46231

Authored-by: Gang Wu <ustcwg@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>