GH-45608: [C++][Flight] Fix compilation for clang (#46264)
This fixes issue #45608 by just moving the definition of the classes `Ticket`, `Location` and `FlightEndpoint` before `FlightInfo`.

### Rationale for this change
`FlightInfo` class in `types.h` uses `FlightEndpoint` in a vector, and clang has problems building this if the definition of `FlightEndpoint` comes after the definition of `FlightInfo`.

### What changes are included in this PR?
By moving the classes a bit up in the file, the problem is resolved.

### Are these changes tested?
Compilation (with GCC) still works, and including the results in another project that is compiled with clang is now successful.

### Are there any user-facing changes?
Not really. The only change is that it should work for clang now 🥳

* GitHub Issue: #45608

Authored-by: Roman Karlstetter <Roman.Karlstetter@ifta.com>
Signed-off-by: David Li <li.davidm96@gmail.com>