GH-46157: [C++] Move test utility RunEndEncodeTableColumns that uses REE to test_util_internal on acero instead of common gtest_util (#46161)
### Rationale for this change

When compiling without `ARROW_COMPUTE` we should not require the run end encode kernel to be available. The current test utility `RunEndEncodeTableColumns` should not be on gtest_util as this is going to be compiled whether `ARROW_COMPUTE` flag is enabled or not.

### What changes are included in this PR?

Move the `RunEndEncodeTableColumns` utility that requires REE to be available to `arrow/acero/test_util_internal` as is only used there, move corresponding test to its own test module and remove the requirement for it to be exported with `ARROW_TESTING_EXPORT`

### Are these changes tested?

Yes, on CI and tested locally that `ARROW_COMPUTE=ON` and `ARROW_COMPUTE=OFF` behave as expected.

### Are there any user-facing changes?

If someone was using the `RunEndEncodeTableColumns` from `libarrow_testing` this is not being exported anymore.

* GitHub Issue: #46157

Lead-authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Co-authored-by: Antoine Pitrou <pitrou@free.fr>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>