GH-46241: [Release][Packaging] Add support for regenerating metadata of APT repositories (#46277)
### Rationale for this change

In general, we don't need to use `dev/release/binary-recover.sh`. We need this only when we can't detect broken APT repositories in RC. But we must be able to detect it automatically. I'll propose it later.

We may document this when:

1. We add support for regenerating metadata for Yum repositories
2. We can't add a mechanism that detects broken APT/Yum repositories

### What changes are included in this PR?

1. Download all `.deb` (and related) files
2. Regenerate metadata of APT repositories from the downloaded files
3. Upload regenerated metadata of APT repositories

### Are these changes tested?

Yes. I fixed our broken APT repositories for 20.0.0 by this.

### Are there any user-facing changes?

No.
* GitHub Issue: #46241

Authored-by: Sutou Kouhei <kou@clear-code.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>