GH-46193: [Flight][Format] Extend Flight Location URI Semantics (#46194)
### Rationale for this change
Updating the documentation in Flight.proto and Flight.rst to extend the
semantics of the allowed Flight location URIs.

### What changes are included in this PR?
Just documentation changes. Currently, none of the Arrow Flight
implementations actually implement handling of the returned URIs beyond
possibly parsing them and wrapping in a `Location` structure. It is left
to the consumer to implement the logic of whether to re-use the same
client or spin up a new client with the new location etc. to perform the
`DoGet` request against. As such, there wasn't a need to make any
code/library changes to accomodate this as part of this PR.


* GitHub Issue: #46193

---------

Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Co-authored-by: Ian Cook <ianmcook@gmail.com>
Co-authored-by: Raúl Cumplido <raulcumplido@gmail.com>