GH-46355: [Python] Fix table.to_struct_array with an empty table (#46357)
### Rationale for this change

Currently `pyarrow.Table.to_struct_array` fails if table is empty, this PR fixes that. See issue #46355 for the reproducible example.

### What changes are included in this PR?

Check if table is empty and create an empty chunked array using table schema for the array type

### Are these changes tested?

I added a test for that, it failed before I implemented the fix

### Are there any user-facing changes?

No
* GitHub Issue: #46355

Authored-by: Konstantin Malanchev <hombit@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>