GH-46222: [Python] Allow to specify footer metadata when opening IPC file for writing (#46354)
### Rationale for this change
The Arrow APIs for other languages offer to set the footer metadata of an IPC file, but `pyarrow` was lacking this. I opened #46222 because we need this feature at KNIME, and took a shot at suggesting an implementation. Happy to get your feedback on the change!

### What changes are included in this PR?
A new keyword argument `metadata` in `RecordBatchFileWriter.__init__` as well as in `ipc.new_file`. The value is `None` by default to not break backwards compatibility.

Also added a `metadata` property to the `RecordBatchFileReader` to be able to extract the metadata easily.

### Are these changes tested?
Yes, by a unit test.

### Are there any user-facing changes?
Yes, see above :)
* GitHub Issue: #46222

Lead-authored-by: Carsten Haubold <carsten.haubold@knime.com>
Co-authored-by: Carsten Haubold <CarstenHaubold@googlemail.com>
Co-authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Rok Mihevc <rok@mihevc.org>