Fix docs for ConcatenateTables+PromoteTableToSchema
The grouping operator for ConcatenateTables was unclosed which caused issues. I just removed the group since we weren't using it. It's possible to fix it without removing the group but I figured it out once and could re-figure-it-out so I just did this.

Once the above was fixed, PromoteTableToSchema started throwing docs warnings (because it could now see the functions) and I fixed this by specifying the variants explicitly.

This now looks right locally.