GH-46373: [Python] Exercise fallback case on tests for parquet.read_table in case dataset is not available (#46550)
### Rationale for this change

Since we merged:
- https://github.com/apache/arrow/pull/39112

The fallback case were dataset is not available is not being tested on parquet.read_table (I did validate myself with a breakpoint).

### What changes are included in this PR?

Adds a basic test that checks the exercises that code path and tests the read_table method from `ParquetFile` instead of `ParquetDataset` in case dataset is not available.

### Are these changes tested?

Yes via CI.

### Are there any user-facing changes?

No

* GitHub Issue: #46373

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Raúl Cumplido <raulcumplido@gmail.com>