GH-43623: [R] remove libarrow backwards compatibility enforcement (#46491)
### Rationale for this change
Remove checks and enforcement of minimum versions. See discussion in the issue. This caused more confusion and discussion than it was helpful. We have the CI in git history if we need it.

### What changes are included in this PR?

Removing the CI job

### Are these changes tested?

They are the tests

### Are there any user-facing changes?

No, we are still technically compatible with 19.0.1, but we don't explicitly support that or use it anywhere. The R package version should match the libarrow version to ensure compatibility — using anything else might work but is off label.
* GitHub Issue: #43623

Authored-by: Jonathan Keane <jkeane@gmail.com>
Signed-off-by: Jonathan Keane <jkeane@gmail.com>