GH-46462: [C++][Parquet] Expose currently thrown EncodedStatistics when checking is_stats_set (#46463)
### Rationale for this change

We are currently throwing away `EncodedStatistics` that are built when calling `is_stats_set`. If we require `EncodedStatistics` we are doing some spurious conversions that could potentially be optimized. For example generating the `TypedStatistics` only when necessary.

### What changes are included in this PR?

- cache the generated `encoded_statistics` on `is_stats_set`
- provide accessor to `encoded_statistics`
- avoid generating the typed statistics if all we care is about `encoded_statistics`

### Are these changes tested?

Existing tests are successful and adapted tests to validate new accessor and `encoded_statistics` values.

### Are there any user-facing changes?

No

* GitHub Issue: #46462

Lead-authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Co-authored-by: Antoine Pitrou <pitrou@free.fr>
Signed-off-by: Antoine Pitrou <antoine@python.org>