GH-46080 [Python][Docs] Provide guidance for tzdata related issues if installing with pip (#46591)
Thanks for opening a pull request!

### Rationale for this change

When installing pyarrow with pip, it is possible to experience a confusing time zone related error when writing files with datetime data.    #46080 discusses the problem and a solution was eventually found.  This change provides helpful information to other users who may experience the problem.

### What changes are included in this PR?

Update to python installation instructions that offers guidance on how to handle tzdata related errors

### Are these changes tested?

I was unable to build the documentation locally.  I will use the github action to preview the documentation

### Are there any user-facing changes?

No

* GitHub Issue: #46080

Lead-authored-by: Kyle Hemker <k@hemker.us>
Co-authored-by: Alenka Frim <AlenkaF@users.noreply.github.com>
Co-authored-by: kahemker <github@kylehemker.uk>
Co-authored-by: kahemker <kyle.hemker@gmail.com>
Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Co-authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: AlenkaF <frim.alenka@gmail.com>