GH-46623: [C++][Compute] Fix the failure of large memory test in arrow-compute-row-test (#46635)
### Rationale for this change

In #45336 we refined the row table buffer accessors and enforced the validation on who can call the `var_length_rows()` method. However a legacy test `CompareColumnsToRowsOver4GBFixedLength` is leveraging this accessor to assert this buffer being null.

### What changes are included in this PR?

We can just check if the row table is fixed length.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

None.

* GitHub Issue: #46623

Authored-by: Rossi Sun <zanmato1984@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>