GH-46219: [C++][Parquet] Remove PARQUET_MINIMAL_DEPENDENCY option (#46274)
### Rationale for this change

This option has been in our code base for some time but is not tested and may no longer work. The consensus in https://github.com/apache/arrow/issues/46219 was to remove it.

### What changes are included in this PR?

The option is removed!

### Are these changes tested?

Yes (covered by the Parquet builds that had previously not tested this option!)

### Are there any user-facing changes?

Yes, a build option was removed; however, I wasn't able to find where this build option was documented in the first place!

* GitHub Issue: #46219

Authored-by: Dewey Dunnington <dewey@wherobots.com>
Signed-off-by: Dewey Dunnington <dewey@wherobots.com>