MINOR: [R][Docs] Fix package title text not showing (#46640)
### Rationale for this change

On the current docs, the title of the package doesn't show up. It would be good if it did. I tried various sanctioned things but none worked. I am chalking it up as something to do with the cosmo theme we're using and a limitation of pkgdown's customization ability.

![image](https://github.com/user-attachments/assets/ffbc2540-f11e-4e68-8ac6-2e6de4ce8b70)

### What changes are included in this PR?

pkgdown will now insert a hard-coded CSS style to make the package text show up in white.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

No.

Authored-by: Bryce Mecum <petridish@gmail.com>
Signed-off-by: Bryce Mecum <petridish@gmail.com>