GH-46508: [C++] Upgrade OpenTelemetry cpp to avoid build error on recent Clang (#46509)
### Rationale for this change

A warning is introduced in Clang 19.1.0 and AppleClang 17.0.0 that won't be demoted by `-Wno-error`, causing opentelemetry-cpp build failure.

### What changes are included in this PR?

Upgrade opentelemetry-cpp (and opentelemetry-proto correspondingly) to the most recent version, which has addressed this issue in https://github.com/open-telemetry/opentelemetry-cpp/pull/3133.

With this upgrade, we found several false-positives in sanitizers. The reason seems to be that we build bundled third-party dependencies as static libraries and don't instrument their code. This is known to cause false-positives as per https://github.com/google/sanitizers/wiki/threadsanitizercppmanual#non-instrumented-code . So some suppressions and disablements are also made in this PR.

### Are these changes tested?

Manually build pass.

### Are there any user-facing changes?

None.
* GitHub Issue: #46508

Authored-by: Rossi Sun <zanmato1984@gmail.com>
Signed-off-by: Rossi Sun <zanmato1984@gmail.com>