GH-46439: [C++] Rename internal Converter class in from_string.cc (#46697)
### Rationale for this change

In https://github.com/apache/arrow/issues/45908 we moved the Converter class in from_string.cc to the internal namespace to avoid a symbol clash with the Converter class defined in arrow/util. It's better to keep the class in an anonymous namespace since it's internal to the file. This reverts the previous change and just renames the class.

### What changes are included in this PR?

- Removed namespace, use anonymous namespace instead like before
- Renamed from Converter to JSONConverter

### Are these changes tested?

Yes.

### Are there any user-facing changes?

No. These changes are internal-only.
* GitHub Issue: #46439

Authored-by: Bryce Mecum <petridish@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>